+++ src/main/kotlin/br/com/pradofigu/maestro/input/restapi/product/dto/ProductRequest.kt
@@ -2,14 +2,14 @@ package br.com.pradofigu.maestro.input.restapi.product.dto
 
 import br.com.pradofigu.maestro.domain.product.model.Product
 import br.com.pradofigu.maestro.input.restapi.category.dto.CategoryRequest
-import reactor.kotlin.core.publisher.toMono
 import java.math.BigDecimal
+import java.math.BigInteger
 
 data class ProductRequest(
     val name: String,
     val price: BigDecimal,
     val category: CategoryRequest,
-    val preparationTime: BigDecimal
+    val preparationTime: BigInteger
 ) {
     fun toModel() = Product(
         name = name,
